Refactoring cartelle di test
